--- a/Common/SecurityIdentifier.cs
+++ b/Common/SecurityIdentifier.cs
@@ -98,8 +98,7 @@ namespace QuantConnect
         private readonly string _symbol;
         private readonly ulong _properties;
         private readonly SidBox _underlying;
-        private readonly Lazy<SecurityType> _lazySecurityType;
-        private readonly Lazy<int> _lazyHashCode;
+        private readonly int _hashCode;
 
         #endregion
 
@@ -186,7 +185,7 @@ namespace QuantConnect
         /// <summary>
         /// Gets the security type component of this security identifier.
         /// </summary>
-        public SecurityType SecurityType => _lazySecurityType.Value;
+        public SecurityType SecurityType { get; }
 
         /// <summary>
         /// Gets the option strike price. This only applies to SecurityType.Option
@@ -264,8 +263,8 @@ namespace QuantConnect
             _symbol = symbol;
             _properties = properties;
             _underlying = null;
-            _lazySecurityType = new Lazy<SecurityType>(() => (SecurityType)ExtractFromProperties(SecurityTypeOffset, SecurityTypeWidth, properties));
-            _lazyHashCode = new Lazy<int>(() => unchecked (symbol.GetHashCode() * 397) ^ properties.GetHashCode());
+            SecurityType = (SecurityType)ExtractFromProperties(SecurityTypeOffset, SecurityTypeWidth, properties);
+            _hashCode = unchecked (symbol.GetHashCode() * 397) ^ properties.GetHashCode();
         }
 
         /// <summary>
@@ -724,7 +723,7 @@ namespace QuantConnect
         /// A hash code for the current <see cref="T:System.Object"/>.
         /// </returns>
         /// <filterpriority>2</filterpriority>
-        public override int GetHashCode() => _lazyHashCode.Value;
+        public override int GetHashCode() => _hashCode;
 
         /// <summary>
         /// Override equals operator